Safari-609.1.11.1/LayoutTests/crypto/subtle/ecdsa-generate-export-key-pkcs8.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
<script src="../resources/common.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>

<script type="text/javascript">
description("Test exporting an ECDSA private key with PKCS8 format.");

jsTestIsAsync = true;

var algorithmKeyGen = {
    name: "ECDSA",
    namedCurve: "P-256"
};
var extractable = true;

var keyPair;
debug("Generating a key pair...");
crypto.subtle.generateKey(algorithmKeyGen, extractable, ["sign", "verify"]).then(function(result) {
    keyPair = result;
    debug("Exporting the public key...");
    return crypto.subtle.exportKey("pkcs8", keyPair.privateKey);
}).then(function(result) {
    privateKey = result;

    shouldBe("privateKey.byteLength", "138");

    finishJSTest();
});
</script>

<script src="../../resources/js-test-post.js"></script>
</body>
</html>
