LayoutTests/crypto/subtle/ecdsa-generate-export-key-pkcs8-expected.txt

Test exporting an ECDSA private key with PKCS8 format.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Generating a key pair...
Exporting the public key...
PASS privateKey.byteLength is 138
PASS successfullyParsed is true

TEST COMPLETE
